GH-101: Rename unnamed visualization to InteractiveViewer. Remove old GraphExplorer and InteractiveViewer/Visualization code, tests, and examples.
