+++ src/ipyradiant/__init__.py
@@ -13,27 +13,27 @@ from .basic_tools import (
 from .loader import FileManager, PathLoader, UpLoader
 from .query import QueryWidget, service_patch_rdflib
 from .visualization import (
+    CytoscapeViewer,
     CytoscapeVisualizer,
     DatashaderVisualizer,
-    GetOutgoingPredicateObjects,
     InteractiveViewer,
     LayoutSelector,
 )
 
 __all__ = [
     "__version__",
+    "collapse_predicates",
+    "CustomURIRef",
+    "CytoscapeViewer",
     "CytoscapeVisualizer",
     "DatashaderVisualizer",
     "FileManager",
-    "LayoutSelector",
-    "PathLoader",
-    "QueryWidget",
-    "UpLoader",
-    "MultiPanelSelect",
-    "service_patch_rdflib",
-    "CustomURIRef",
-    "PredicateMultiselectApp",
-    "collapse_predicates",
     "InteractiveViewer",
-    "GetOutgoingPredicateObjects",
+    "LayoutSelector",
+    "MultiPanelSelect",
+    "PathLoader",
+    "PredicateMultiselectApp",
+    "QueryWidget",
+    "service_patch_rdflib",
+    "UpLoader",
 ]
