+++ src/ipyradiant/visualization/cytoscape/__init__.py
@@ -1,7 +1,8 @@
 # Copyright (c) 2021 ipyradiant contributors.
 # Distributed under the terms of the Modified BSD License.
 
-__all__ = ["CytoscapeViewer", "CytoscapeVisualizer"]
+__all__ = ["CytoscapeViewer", "CytoscapeVisualizer", "InteractiveViewer"]
 
 from .app import CytoscapeVisualizer
+from .interactive import InteractiveViewer
 from .viewer import CytoscapeViewer
