+++ src/ipyradiant/visualization/cytoscape/interactive.py
@@ -13,7 +13,6 @@ from .viewer import CytoscapeViewer
 from ipyradiant.query.api import SPARQLQueryFramer
 from ipyradiant.basic_tools.custom_uri_ref import CustomURIRef
 from ipyradiant.visualization.cytoscape import style
-from ipyradiant.visualization.explore.interactive_exploration import add_cyto_class, remove_cyto_class
 
 
 # Ranked by pairwise distance (generated by https://mokole.com/palette.html)
@@ -76,7 +75,7 @@ class TypeCount(SPARQLQueryFramer):
     """
 
 
-class UnnamedVis(W.VBox):
+class InteractiveViewer(W.VBox):
     """TODO
 
     TODO document how users can extend the type_count_query
@@ -134,7 +133,7 @@ class UnnamedVis(W.VBox):
         widget = CytoscapeViewer()
         T.link((self, "graph"), (widget, "graph"))
         widget.cytoscape_widget.on("node", "click", self.load_json)
-        widget.layout.width = "80%"
+        widget.layout.width = "85%"
         return widget
 
     @T.default("type_selector")
@@ -142,16 +141,10 @@ class UnnamedVis(W.VBox):
         widget = W.SelectMultiple(
             layout=W.Layout(
                 border='solid 1px',
-                flex='auto'
             )
         )
         widget.observe(self.update_classes, "value")
         return widget
-
-    # @T.default("json_output")
-    # def _make_default_json_output(self):
-    #     widget = W.Output()
-    #     return widget
 
     @T.validate("children")
     def validate_children(self, proposal):
@@ -163,11 +156,12 @@ class UnnamedVis(W.VBox):
         if not children:
             children = (
                 W.HBox(
-                    [self.type_selector, self.viewer],
-                    layout=W.Layout(height="80%")
+                    [self.type_selector, self.viewer]
                 ),
                 self.json_output,
             )
+
+        # TODO self.layout ?
         return children
 
     @T.observe("graph")
