--- src/ipyradiant/visualization/explore/__init__.py
@@ -1,25 +1,0 @@
-# Copyright (c) 2021 ipyradiant contributors.
-# Distributed under the terms of the Modified BSD License.
-
-__all__ = [
-    "GetOutgoingPredicateObjects",
-    "GraphExplorer",
-    "GraphExploreSelect",
-    "GraphExploreSelectMultiple",
-    "InteractiveViewer",
-    "RDFSubjectSelect",
-    "RDFSubjectSelectMultiple",
-    "RDFTypeSelect",
-    "RDFTypeSelectMultiple",
-]
-
-from .graph_explorer import (
-    GraphExplorer,
-    GraphExploreSelect,
-    GraphExploreSelectMultiple,
-    RDFSubjectSelect,
-    RDFSubjectSelectMultiple,
-    RDFTypeSelect,
-    RDFTypeSelectMultiple,
-)
-from .interactive_exploration import GetOutgoingPredicateObjects, InteractiveViewer
